refactored code structure
